move transformer encoder layer to own file
