+++ metaseq/modules/__init__.py
@@ -11,7 +11,8 @@ from .learned_positional_embedding import LearnedPositionalEmbedding
 from .multihead_attention import MultiheadAttention
 from .positional_embedding import PositionalEmbedding
 from .sinusoidal_positional_embedding import SinusoidalPositionalEmbedding
-from .transformer_layer import TransformerDecoderLayer, TransformerEncoderLayer
+from .transformer_layer import TransformerDecoderLayer
+from .transformer_encoder_layer import TransformerEncoderLayer
 from .linear import Linear
 
 __all__ = [
@@ -25,6 +26,5 @@ __all__ = [
     "PositionalEmbedding",
     "SinusoidalPositionalEmbedding",
     "TransformerDecoderLayer",
-    "TransformerEncoderLayer",
     "Linear",
 ]
